rename gen_* files
